# Conflicts:
#	app/src/main/res/layout/tarjetas_gastos.xml

+++ app/src/main/java/com/example/elTrioDeNos/dataClases/Ingreso.kt
@@ -3,7 +3,7 @@ package com.example.elTrioDeNos.dataClases
 import java.util.Date
 
 data class Ingreso(
-    val monto: Int,
+    val monto: Double,
     val nota: String,
     val fecha: Date
 )
